requirements.txt
Fabric3==1.14.post1
python-digitalocean==1.13.2
